+++ serverless.yml
@@ -65,91 +65,6 @@ functions:
         - test/**
         - .git/**
         - .travis.yml
-  transaction:
-    handler: src/lambdas/transaction/transactionLambda.handle
-    events:
-      - http:
-          path: account/{accountId}/wallet/{walletId}/transaction
-          method: GET
-          authorizer:
-            arn: ${self:custom.cognitoIdpArn}
-            scopes:
-              - email
-          parameters:
-            paths:
-              accountId: true
-              walletId: true
-      - http:
-          path: account/{accountId}/wallet/{walletId}/transaction
-          method: PUT
-          authorizer:
-            arn: ${self:custom.cognitoIdpArn}
-            scopes:
-              - email
-          parameters:
-            paths:
-              accountId: true
-              walletId: true
-            querystrings:
-              from: false
-              to: false
-      - http:
-          path: account/{accountId}/wallet/{walletId}/transaction
-          method: DELETE
-          authorizer:
-            arn: ${self:custom.cognitoIdpArn}
-            scopes:
-              - email
-          parameters:
-            paths:
-              accountId: true
-              walletId: true
-      - http:
-          path: account/{accountId}/wallet/{walletId}/transaction/{transactionId}
-          method: GET
-          authorizer:
-            arn: ${self:custom.cognitoIdpArn}
-            scopes:
-              - email
-          parameters:
-            paths:
-              accountId: true
-              walletId: true
-              transactionId: true
-      - http:
-          path: account/{accountId}/wallet/{walletId}/transaction/{transactionId}
-          method: PUT
-          authorizer:
-            arn: ${self:custom.cognitoIdpArn}
-            scopes:
-              - email
-          parameters:
-            paths:
-              accountId: true
-              walletId: true
-              transactionId: true
-      - http:
-          path: account/{accountId}/wallet/{walletId}/transaction/{transactionId}
-          method: DELETE
-          authorizer:
-            arn: ${self:custom.cognitoIdpArn}
-            scopes:
-              - email
-          parameters:
-            paths:
-              accountId: true
-              walletId: true
-              transactionId: true
-    package:
-      include:
-        - src/lambdas/transaction/transactionLambda.js
-        - src/lambdas/transaction/transactionHandler.js
-        - libs/*
-        - node_modules/**
-      exclude:
-        - test/**
-        - .git/**
-        - .travis.yml
   fileRouter:
     handler: src/lambdas/file-router/fileRouterLambda.handle
     environment:
