+++ services/transaction/src/transactionHandler.js
@@ -1,6 +1,6 @@
-const transaction = require('../../libs/transaction');
+const transaction = require('libs/transaction');
 const Transaction = transaction.Transaction;
-const dynamodb = require('../../libs/dynamodb');
+const dynamodb = require('libs/dynamodb');
 const DynamoDb = dynamodb.DynamoDb;
 const QueryBuilder = dynamodb.QueryBuilder;
 const ExpressionBuilder = dynamodb.ExpressionBuilder;
